requirements.txt
requests==2.32.5
python-telegram-bot==13.15
flask==2.3.3